// generated/prisma/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EtiquetaStatus = {
  PENDIENTE: 'PENDIENTE',
  IMPRESA: 'IMPRESA',
  EN_CAMINO: 'EN_CAMINO',
  ENTREGADA: 'ENTREGADA',
  NO_ENTREGADA: 'NO_ENTREGADA'
} as const

export type EtiquetaStatus = (typeof EtiquetaStatus)[keyof typeof EtiquetaStatus]


export const SocialPlatformEnum = {
  facebook: 'facebook',
  instagram: 'instagram',
  whatsapp: 'whatsapp'
} as const

export type SocialPlatformEnum = (typeof SocialPlatformEnum)[keyof typeof SocialPlatformEnum]


export const OrderStatusEnum = {
  PENDIENTE: 'PENDIENTE',
  EN_CURSO: 'EN_CURSO',
  ENTREGADO: 'ENTREGADO',
  CANCELADO: 'CANCELADO'
} as const

export type OrderStatusEnum = (typeof OrderStatusEnum)[keyof typeof OrderStatusEnum]


export const ServiceTypeEnum = {
  EXPRESS: 'EXPRESS',
  LOW_COST: 'LOW_COST',
  PUNTO_DE_RETIRO: 'PUNTO_DE_RETIRO'
} as const

export type ServiceTypeEnum = (typeof ServiceTypeEnum)[keyof typeof ServiceTypeEnum]